+++ pyproject.toml
@@ -1,17 +1,27 @@
+[build-system]
+requires = ["hatchling>=1.24"]
+build-backend = "hatchling.build"
+
 [project]
 name = "telebot-lite"
 version = "0.1.0"
-descripttion = "A lightweight async wrapper for Telegram Bot API 9.0"
+description = "A lightweight async wrapper for Telegram Bot API 9.0"
 requires-python = ">=3.9"
-authors = [{name = "Alireza Sarami", email = "alirs1997@gmail.com"}]
+authors = [{ name = "Alireza Sarami", email = "alirs1997@gmail.com" }]
+
+# لیست وابستگی‌ها (فرمت PEP 508، بدون caret ^)
+dependencies = [
+  "httpx>=0.27,<0.28",
+  "pydantic>=2.8,<3.0"
+]
 
 [project.urls]
-"Homepage" = "https://github.com/alireza_sarami/telebot-lite"
+Homepage = "https://github.com/alireza_sarami/telebot-lite"
 
-[project.dependencies]
-httpx = "^0.27.0"
-pydantic = "^2.8"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+# اختیاری: وابستگی‌های محیط توسعه
+[project.optional-dependencies]
+dev = [
+  "pytest",
+  "ruff",
+  "mypy"
+]
