--- telebot/pyproject.toml
@@ -1,17 +1,0 @@
-[project]
-name = "telebot-lite"
-version = "0.1.0"
-description = "A lightweight async wrapper for Telegram Bot API 9.0"
-requires-python = ">=3.9"
-authors = [{name = "Alireza Sarami", email="alirs1997@gmail.com"}]
-
-[project.urls]
-"Homepage" = "https://github.com/alireza_sarami/telebot-lite"
-
-[project.dependencies]
-httpx = "^0.27.0"
-pydantic = "^2.8"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
