+++ telebot_lite/core/exception.py
